stockwatch/requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
blessed==1.20.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.5
django-celery-results==2.5.1
django-picklefield==3.2
django-q2==1.7.6
django-timezone-field==7.1
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==3.5.3
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
